メソッドをDecoratorからModelに移動
