+++ app/models/practice.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Practice < ApplicationRecord
+  OGP_IMAGE_SIZE = [1200, 630].freeze
+
   include Watchable
   include Searchable
 
@@ -166,6 +168,18 @@ class Practice < ApplicationRecord
     return false if practices_books.empty?
 
     practices_books.any?(&:must_read)
+  end
+
+  def ogp_image_url
+    default_image_path = '/ogp/ogp.png'
+
+    if ogp_image.attached?
+      ogp_image.variant(resize_to_fill: OGP_IMAGE_SIZE, autorot: true, saver: { strip: true, quality: 60 }).processed.url
+    else
+      image_url default_image_path
+    end
+  rescue ActiveStorage::FileNotFoundError, ActiveStorage::InvariableError, Vips::Error
+    image_url default_image_path
   end
 
   private
